Refactor code
Split code for initializing and running the simulation algorithm into class 'Algorithm',  and the code for rendering the particles into class 'ParticleRender'.
Create '.env' for storing the NUL value.
Refactor 'ParticleGeometry' class to be more readable.
Change 'GPUCompute' to be able to work off a reference to an updating array, and instead of using the varInputs Proxy setter, for the individual varyings, just call the 'updateVaryings()' method.